# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: Learning Assistant API
  description: |
    Comprehensive API for the Learning Assistant platform - an intelligent, adaptive learning system that personalizes education based on individual learning styles, behavior patterns, and performance metrics.
    
    ## Key Features
    - **Adaptive Learning**: Real-time adjustment of content delivery based on user performance
    - **VARK Learning Styles**: Support for Visual, Auditory, Reading/Writing, and Kinesthetic preferences
    - **AI-Powered Recommendations**: Intelligent content suggestions and study paths
    - **Real-time Analytics**: Performance tracking and learning insights
    - **Security-First**: Enterprise-grade security with RBAC, rate limiting, and comprehensive monitoring
    
    ## Authentication
    This API uses JWT Bearer token authentication with role-based access control (RBAC).
    Include the token in the Authorization header: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    API endpoints have different rate limits based on the endpoint type and user role:
    - General endpoints: 1000 requests/hour per user, 100 requests/minute per IP
    - Learning endpoints: 200 requests/minute per user
    - Administrative endpoints: 50 requests/minute per admin
    
    ## Security Features
    - CSRF protection for state-changing operations
    - Input validation and sanitization
    - Comprehensive audit logging
    - Real-time threat detection
    
  version: 1.0.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get Analytics Dashboard Data
      description: |
        Retrieve comprehensive analytics data for the dashboard.
        Includes performance metrics, user engagement, and system health.
        Admin access required.
      operationId: getAnalyticsDashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Analytics data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsDashboardResponse'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (admin required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Security Operations
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication. Include the token in the Authorization header.
        Tokens are issued upon successful authentication and have a configurable expiration time.
  schemas:
...
    PerformanceMetrics:
      type: object
      properties:
        requests:
          type: integer
          description: Total number of requests processed
        errors:
          type: integer
          description: Number of errors encountered
        slowRequests:
          type: integer
          description: Number of slow requests
        averageResponseTime:
          type: integer
          description: Average response time in milliseconds
        activeTraces:
          type: integer
          description: Number of active traces
        throughput:
          type: number
          description: Requests per second
...
    AnalyticsDashboardResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          description: Indicates successful operation
        data:
          type: object
          properties:
            userMetrics:
              $ref: '#/components/schemas/UserMetrics'
            performanceMetrics:
              $ref: '#/components/schemas/PerformanceMetrics'
            learningMetrics:
              $ref: '#/components/schemas/LearningMetrics'
            systemMetrics:
              $ref: '#/components/schemas/SystemMetrics'
    UserMetrics:
      type: object
      properties:
        totalUsers:
          type: integer
          description: Total number of registered users
        activeUsers:
          type: integer
          description: Number of users active in the last 30 days
        newUsers:
          type: integer
          description: Number of new users in the last 30 days
        userGrowthRate:
          type: number
          description: User growth rate percentage
    LearningMetrics:
      type: object
      properties:
        totalSessions:
          type: integer
          description: Total number of learning sessions
        averageSessionDuration:
          type: number
          description: Average session duration in minutes
        completionRate:
          type: number
          minimum: 0
          maximum: 100
          description: Overall completion rate percentage
        averageScore:
          type: number
          minimum: 0
          maximum: 100
          description: Average score across all assessments
    SystemMetrics:
      type: object
      properties:
        uptime:
          type: number
          description: System uptime percentage
        errorRate:
          type: number
          description: Error rate percentage
        averageResponseTime:
          type: number
          description: Average response time in milliseconds
        throughput:
          type: number
          description: Requests per second

    # Security Schemas
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type or code
        message:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        details:
          type: object
          description: Additional error details
        requestId:
          type: string
          description: Unique request identifier for debugging
        statusCode:
          type: integer
          description: HTTP status code
...